+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+
+## [1.1.0] - 2020-03-06
 ### Added
 
 - Support for the [new Apple Podcast categories][category-new-2019] that were [added August 9th 2019][category-published-2019].
@@ -51,6 +54,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The Podcast and Episode classes for easily generating a podcast out of data,
   and related utilities and classes.
 
-[Unreleased]: https://github.com/tobinus/python-podgen/compare/v1.0.1...develop
+[Unreleased]: https://github.com/tobinus/python-podgen/compare/v1.1.0...develop
+[1.1.0]: https://github.com/tobinus/python-podgen/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/tobinus/python-podgen/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/tobinus/python-podgen/compare/290045ac...v1.0.0